+++ src/main/gen/com/tbusk/vala_plugin/psi/ValaArgument.java
@@ -11,9 +11,6 @@ public interface ValaArgument extends PsiElement {
   ValaExpression getExpression();
 
   @Nullable
-  ValaValidIdentifierKeywords getValidIdentifierKeywords();
-
-  @Nullable
-  PsiElement getIdentifier();
+  ValaIdentifier getIdentifier();
 
 }