+++ src/main/gen/com/tbusk/vala_plugin/psi/ValaEnumvalue.java
@@ -13,10 +13,7 @@ public interface ValaEnumvalue extends PsiElement {
   @Nullable
   ValaExpression getExpression();
 
-  @Nullable
-  ValaValidIdentifierKeywords getValidIdentifierKeywords();
-
-  @Nullable
-  PsiElement getIdentifier();
+  @NotNull
+  ValaIdentifier getIdentifier();
 
 }