+++ src/main/gen/com/tbusk/vala_plugin/psi/ValaTypeParameters.java
@@ -8,6 +8,6 @@ import com.intellij.psi.PsiElement;
 public interface ValaTypeParameters extends PsiElement {
 
   @NotNull
-  List<ValaValidIdentifierKeywords> getValidIdentifierKeywordsList();
+  List<ValaIdentifier> getIdentifierList();
 
 }