+++ src/main/gen/com/tbusk/vala_plugin/psi/impl/ValaFieldDeclarationSectionImpl.java
@@ -8,12 +8,13 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.tbusk.vala_plugin.psi.ValaTypes.*;
-import com.intellij.extapi.psi.ASTWrapperPsiElement;
+import com.tbusk.vala_plugin.psi.ValaNamedElementImpl;
 import com.tbusk.vala_plugin.psi.*;
+import com.intellij.navigation.ItemPresentation;
 
-public class ValaFieldDeclarationSectionImpl extends ASTWrapperPsiElement implements ValaFieldDeclarationSection {
+public class ValaFieldDeclarationSectionImpl extends ValaNamedElementImpl implements ValaFieldDeclarationSection {
 
-  public ValaFieldDeclarationSectionImpl(@NotNull ASTNode node) {
+  public ValaFieldDeclarationSectionImpl(@Nullable ASTNode node) {
     super(node);
   }
 
@@ -35,20 +36,34 @@ public class ValaFieldDeclarationSectionImpl extends ASTWrapperPsiElement implem
 
   @Override
   @NotNull
+  public ValaIdentifier getIdentifier() {
+    return findNotNullChildByClass(ValaIdentifier.class);
+  }
+
+  @Override
+  @NotNull
   public List<ValaInlineArrayType> getInlineArrayTypeList() {
     return PsiTreeUtil.getChildrenOfTypeAsList(this, ValaInlineArrayType.class);
   }
 
   @Override
-  @Nullable
-  public ValaValidIdentifierKeywords getValidIdentifierKeywords() {
-    return findChildByClass(ValaValidIdentifierKeywords.class);
+  public String getName() {
+    return ValaPsiImplUtil.getName(this);
   }
 
   @Override
-  @Nullable
-  public PsiElement getIdentifier() {
-    return findChildByType(IDENTIFIER);
+  public PsiElement setName(String newName) {
+    return ValaPsiImplUtil.setName(this, newName);
+  }
+
+  @Override
+  public PsiElement getNameIdentifier() {
+    return ValaPsiImplUtil.getNameIdentifier(this);
+  }
+
+  @Override
+  public ItemPresentation getPresentation() {
+    return ValaPsiImplUtil.getPresentation(this);
   }
 
 }