+++ src/main/gen/com/tbusk/vala_plugin/psi/impl/ValaNamespaceDeclarationImpl.java
@@ -8,12 +8,13 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.tbusk.vala_plugin.psi.ValaTypes.*;
-import com.intellij.extapi.psi.ASTWrapperPsiElement;
+import com.tbusk.vala_plugin.psi.ValaNamedElementImpl;
 import com.tbusk.vala_plugin.psi.*;
+import com.intellij.navigation.ItemPresentation;
 
-public class ValaNamespaceDeclarationImpl extends ASTWrapperPsiElement implements ValaNamespaceDeclaration {
+public class ValaNamespaceDeclarationImpl extends ValaNamedElementImpl implements ValaNamespaceDeclaration {
 
-  public ValaNamespaceDeclarationImpl(@NotNull ASTNode node) {
+  public ValaNamespaceDeclarationImpl(@Nullable ASTNode node) {
     super(node);
   }
 
@@ -43,6 +44,26 @@ public class ValaNamespaceDeclarationImpl extends ASTWrapperPsiElement implement
   @NotNull
   public List<ValaUsingDirective> getUsingDirectiveList() {
     return PsiTreeUtil.getChildrenOfTypeAsList(this, ValaUsingDirective.class);
+  }
+
+  @Override
+  public String getName() {
+    return ValaPsiImplUtil.getName(this);
+  }
+
+  @Override
+  public PsiElement setName(String newName) {
+    return ValaPsiImplUtil.setName(this, newName);
+  }
+
+  @Override
+  public PsiElement getNameIdentifier() {
+    return ValaPsiImplUtil.getNameIdentifier(this);
+  }
+
+  @Override
+  public ItemPresentation getPresentation() {
+    return ValaPsiImplUtil.getPresentation(this);
   }
 
 }