+++ src/main/gen/com/tbusk/vala_plugin/psi/impl/ValaTypeImpl.java
@@ -35,12 +35,6 @@ public class ValaTypeImpl extends ASTWrapperPsiElement implements ValaType {
 
   @Override
   @Nullable
-  public ValaPrimitiveType getPrimitiveType() {
-    return findChildByClass(ValaPrimitiveType.class);
-  }
-
-  @Override
-  @Nullable
   public ValaSymbol getSymbol() {
     return findChildByClass(ValaSymbol.class);
   }