+++ src/main/java/com/tbusk/vala_plugin/highlighting/syntax/ValaErrorDomainDeclarationHighlighter.java
@@ -1,0 +1,38 @@
+package com.tbusk.vala_plugin.highlighting.syntax;
+
+import com.intellij.lang.annotation.AnnotationHolder;
+import com.intellij.psi.PsiElement;
+import com.tbusk.vala_plugin.highlighting.ValaHighlighter;
+import com.tbusk.vala_plugin.highlighting.ValaHighlighterUtil;
+import com.tbusk.vala_plugin.highlighting.ValaSyntaxHighlightingAnnotator;
+import com.tbusk.vala_plugin.highlighting.ValaTextAttributeKey;
+import com.tbusk.vala_plugin.psi.impl.ValaErrordomainDeclarationImpl;
+import org.jetbrains.annotations.NotNull;
+
+public final class ValaErrorDomainDeclarationHighlighter implements ValaHighlighter {
+
+    private static volatile ValaErrorDomainDeclarationHighlighter instance;
+
+    private ValaErrorDomainDeclarationHighlighter() {
+    }
+
+    public static synchronized ValaErrorDomainDeclarationHighlighter getInstance() {
+        if (instance == null) {
+            instance = new ValaErrorDomainDeclarationHighlighter();
+        }
+
+        return instance;
+    }
+
+    public void highlight(@NotNull PsiElement psiElement, @NotNull AnnotationHolder annotationHolder) {
+        if (psiElement instanceof ValaErrordomainDeclarationImpl) {
+
+            ValaSyntaxHighlightingAnnotator.addScopedElement(psiElement);
+
+            ValaHighlighterUtil util = ValaHighlighterUtil.getInstance();
+
+            util.highlightSymbol(psiElement, annotationHolder, ValaTextAttributeKey.STRUCTURE_NAMES);
+        }
+    }
+
+}