+++ src/main/java/com/tbusk/vala_plugin/psi/ValaNamedElement.java
@@ -1,0 +1,7 @@
+package com.tbusk.vala_plugin.psi;
+
+import com.intellij.psi.PsiNameIdentifierOwner;
+import com.intellij.psi.PsiNamedElement;
+
+public interface ValaNamedElement extends PsiNamedElement, PsiNameIdentifierOwner {
+}