+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### What's Changed
 
--
+- Improved syntax highlighting (declarations, identifier, reference, etc.)
 
 ## [1.2.4-ALPHA]